+++ src/ReadoutEquipmentRORC.cxx
@@ -144,13 +144,6 @@ ReadoutEquipmentRORC::ReadoutEquipmentRORC(ConfigFile &cfg, std::string name)
     std::string cfgDataSource = "Internal";
     cfg.getOptionalValue<std::string>(name + ".dataSource", cfgDataSource);
 
-    // configuration parameter: | equipment-rorc-* | linkMask | string | 0-11 |
-    // List of links to be enabled. For CRU, in the 0-11 range. Can be a single
-    // value, a comma-separated list, a range or comma-separated list of ranges.
-    // c.f. AliceO2::roc::Parameters. |
-    std::string cfgLinkMask = "0-11";
-    cfg.getOptionalValue<std::string>(name + ".linkMask", cfgLinkMask);
-
     // std::string cfgReadoutMode="CONTINUOUS";
     // cfg.getOptionalValue<std::string>(name + ".readoutMode", cfgReadoutMode);
 
@@ -272,11 +265,6 @@ ReadoutEquipmentRORC::ReadoutEquipmentRORC(ConfigFile &cfg, std::string name)
     theLog.log("Register DMA block %p:%lu", baseAddress, blockSize);
     params.setBufferParameters(
         AliceO2::roc::buffer_parameters::Memory{baseAddress, blockSize});
-
-    // define link mask
-    // this is harmless for C-RORC
-    params.setLinkMask(
-        AliceO2::roc::Parameters::linkMaskFromString(cfgLinkMask));
 
     // open channel with above parameters
     channel = AliceO2::roc::ChannelFactory().getDmaChannel(params);